UserController User Call

+++ Stories.Data.InitialData/TimelineData.cs
@@ -19,11 +19,19 @@ namespace Stories.Data.InitialData
                 Timeline timeline = new Timeline();
                 timeline.PersonId = Guid.Parse("019520F8-E48B-4079-84CC-B7F0F5A79C1F");
                 timeline.TimelineName = "Naofumi Aoyama";
+                timeline.CreateUserId = Guid.Parse("019520F8-E48B-4079-84CC-B7F0F5A79C1F");
+                timeline.CreateDate = DateTime.Now;
+                timeline.UpdateUserId = Guid.Parse("019520F8-E48B-4079-84CC-B7F0F5A79C1F");
+                timeline.UpdateDate = DateTime.Now;
                 await timelineRepository.Add(timeline);
 
                 Timeline timeline2 = new Timeline();
                 timeline2.PersonId = Guid.Parse("F7A70CB7-F46D-4A94-88CD-6B0284CBE96F");
                 timeline2.TimelineName = "Jenalyn Albios";
+                timeline2.CreateUserId = Guid.Parse("019520F8-E48B-4079-84CC-B7F0F5A79C1F");
+                timeline2.CreateDate = DateTime.Now;
+                timeline2.UpdateUserId = Guid.Parse("019520F8-E48B-4079-84CC-B7F0F5A79C1F");
+                timeline2.UpdateDate = DateTime.Now;
                 await timelineRepository.Add(timeline2);
             }
         }
@@ -37,10 +45,7 @@ namespace Stories.Data.InitialData
                 if (timeline1 != null) { await timelineRepository.Delete(timeline1); }
 
                 var timeline2 = timelineRepository.Get(Guid.Parse("F7A70CB7-F46D-4A94-88CD-6B0284CBE96F")).Result;
-                if (timeline2 != null)
-                {
-                    await timelineRepository.Delete(timeline2);
-                }
+                if (timeline2 != null) { await timelineRepository.Delete(timeline2); }
             }
         }
